Add reindenting for multiple possible indents
Some lines of Grace code can potentially be at multiple indentation
levels. Now lines which could be either a newline or a continuation of a
request on the previous line will default to a newline, but indenting
twice in the same place will push it out to a continuation or back to a
newline. If the line's indentation is already more than the line before,
it will default to a continuation.

This change has exposed some strange behaviour with electric-indent
where the indentation function is invoked twice on a newline, which
triggers the reindent when only an indent is desired. For now,
electric-indent is disabled for grace-mode, but in the future
reindenting could be a separate feature only enabled for a press of TAB.