Fix method header detection and indent calculation
Placing the cursor on an empty line in the middle of a method header
caused the header detection to fail because the statement rewind saw the
empty line as a blank statement. Now the detection rewinds to the last
significant character before continuing header detection.

Checking indentation values inside of parentheses also didn't work
correctly, because it checked the indentation check used the current
paren-level when it only cared about the brackets. Indentation checking
now ignores parentheses, and grace-paren-level has been renamed to the
more accurate grace-list-level.